requirements.txt
torch
torchvision
click
PyYAML
scikit-learn
pandas
Pillow
